refactor: restructure settings components with new organization and exempt settings pages
