+++ src/app/(stacked)/settings/exempts/page.tsx
@@ -1,0 +1,5 @@
+import { ExemptSettings } from "@/components/feature";
+
+export default function ExemptSettingsPage() {
+  return <ExemptSettings />;
+}
