+++ src/app/(stacked)/settings/organizations/page.tsx
@@ -1,0 +1,5 @@
+import { OrganizationSettings } from "@/components/feature";
+
+export default function OrganizationSettingsPage() {
+  return <OrganizationSettings />;
+}
