+++ src/app/(stacked)/settings/profile/page.tsx
@@ -1,0 +1,5 @@
+import { ProfileSettings } from "@/components/feature";
+
+export default function ProfileSettingsPage() {
+  return <ProfileSettings />;
+}
