+++ src/components/feature/settings/organization/index.ts
@@ -1,0 +1,1 @@
+export * from "./organization-settings";
