--- src/components/feature/settings/organizations/index.ts
@@ -1,1 +1,0 @@
-export * from "./organizations-settings";
